SCENARIUSZE_JUTRA/requirements.txt
# Scenariusze Jutra - Wymagania
# Narzędzie do analizy foresightowej dla MSZ

# Podstawowe biblioteki
python>=3.9

# NLP i analiza danych
openai>=1.0.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20
beautifulsoup4>=4.12.0
requests>=2.31.0
newspaper3k>=0.2.8
lxml_html_clean>=0.4.0  # Wymagane dla newspaper3k w Python 3.13+
feedparser>=6.0.10

# Przetwarzanie tekstu
nltk>=3.8.1
spacy>=3.7.0
transformers>=4.35.0
sentence-transformers>=2.2.2

# Analiza danych
pandas>=2.1.0
numpy>=1.24.0
scikit-learn>=1.3.0
networkx>=3.2.0

# Wizualizacja
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.18.0

# Baza danych i pamięć
sqlite3  # wbudowane w Python
pickle  # wbudowane w Python

# Bezpieczeństwo
cryptography>=41.0.0

# Narzędzia pomocnicze
python-dotenv>=1.0.0
tqdm>=4.66.0
pydantic>=2.5.0
tenacity>=8.2.0

# Web scraping
selenium>=4.15.0
webdriver-manager>=4.0.0

# Async operations
aiohttp>=3.9.0
asyncio  # wbudowane w Python

